Split server.js into route component and controller component. Fixed indententation
