+++ server/views/login.ejs
@@ -29,7 +29,7 @@
             <input type="text" id="username" name="username" placeholder="username" required />
         </div>
         <div>
-            <input type="text" id="password" name="password" placeholder="password" required />
+            <input type="password" id="password" name="password" placeholder="password" required />
         </div>
         <div>
             <input type="submit" value="Login" />
