+++ server/views/register.ejs
@@ -21,10 +21,10 @@
             <input type="text" id="username" name="username" placeholder="username" required />
         </div>
         <div>
-            <input type="text" id="password" name="password" placeholder="password" required />
+            <input type="password" id="password" name="password" placeholder="password" required />
         </div>
         <div>
-            <input type="text" id="confirm_password" name="confirm_password" placeholder="confirm password" required />
+            <input type="password" id="confirm_password" name="confirm_password" placeholder="confirm password" required />
         </div>
         <div>
             <input type="submit" value="Register" />
